pyproject.toml
[tool.poetry]
name = "s2j"
version = "0.1.0"
description = ""
authors = ["Roberto Polli <robipolli@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.3.1"
pyshp = "^2.1.0"
ipython = "^7.14.0"
click = "^7.1.2"
pytest = "^5.4.2"
requests = "^2.23.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
